% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{choose_loci}
\alias{choose_loci}
\title{Choose and filter loci with zeros}
\usage{
choose_loci(analysis_path, counts, chr_order, samples_used_to_filter,
  zeros_permitted = 0, output_folder = "filtered_counts",
  counts_Rdata = "loci_counts.Rdata")
}
\arguments{
\item{analysis_path}{The path to the directory in which 
you want the results to be saved.
If the path does not exist, it will be created for you.}

\item{counts}{A data frame with \code{chr},
\code{leftPos} and \code{strand} 
in the first three columns and sample counts in all other columns}

\item{chr_order}{A vector of chromosomes of interest.
e.g. \code{paste0("chr", 1:22)} or 
\code{paste0("chr", c(1:22, "X", "Y"))}}

\item{samples_used_to_filter}{A character vector with sample names
used to filter the counts}

\item{zeros_permitted}{Loci with less than or equal to 
\code{zeros_permitted} in the counts of the 
\code{samples_used_to_filter} are kept.  Default: 0.}

\item{output_folder}{The folder name to store 
the filtered counts (Default: `filtered_counts`).
Not recommended to change this.}

\item{counts_Rdata}{The name of the 
resulting Rdata object (Default: \code{loci_counts.Rdata}). 
Not recommended to change this.}
}
\value{
A counts data frame (after filtering loci with 
zero counts in any of the samples provided)
}
\description{
Given a list of samples, counts data frame 
(with columns for \code{chr}, \code{leftPos} and \code{strand}),
and chromosomes of interest
return a data frame with loci filtered 
(loci with less than or equal to 
\code{zeros_permitted} in the counts of the 
\code{samples_used_to_filter} are kept).
}
\examples{
filtered_counts = choose_loci("/path/to/analysis",
loci_counts,
c("sample_1", "sample_2", "sample_3"))

# see data(loci_counts) for example counts data frame
}
