% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{fit_controls}
\alias{fit_controls}
\title{Fit the expected proportion of reads at each locus using a set of control samples}
\usage{
fit_controls(analysis_path, counts, sample_types, output_folder,
  output_run_name, iterations, burn_in, precision_prior, mean_priors,
  precision_sigma, mean_sigma)
}
\arguments{
\item{analysis_path}{The path to the directory in which 
you want the results to be saved.  If the path does
 not exist, it will be created for you.}

\item{counts}{A data frame of counts with \code{chr},
\code{leftPos} and \code{strand}
in the first three columns and sample counts in all other columns}

\item{sample_types}{A data frame with two columns
The first column should have the title \code{Sample} and contain 
the name of all the samples you wish to use in this analysis 
(this is typically all the muliplexed samples on a sequencing lane).
The second column should have the title \code{Control} and the 
value of each row should be 1 if the sample is a control 
sample or 0 if it is not.}

\item{output_folder}{The folder name to store the fit data
(Default: `fit_controls`). 
This will be created for you if it doesn't exist already.
Not recommended to change this.}

\item{output_run_name}{The name to prefix the output files.
Default: `fit_controls`. Not recommended to change this.}

\item{iterations}{The number of iterations to run the MCMC.
Default: \code{20000}.}

\item{burn_in}{The number of initial iterations to burn. 
Default: \code{5000}.}

\item{precision_prior}{The Gamma parameters for the prior 
on the sample inverse dispersion parameter, parameterised by 
shape and scale (Default: \code{c(1.5, 1000000)}).}

\item{mean_priors}{The Beta prior parameters for the proportion
 of reads at a locus in a control sample.
This is represented by a matrix with 2 rows and L columns 
(where L is the number of rows in your counts file).
The first row represents the alpha shape parameter and 
the second row represents the beta shape parameter.
Default: \code{matrix(c(1,1), nrow=2, ncol=nrow(counts))}
i.e. a flat `Beta(1,1)` prior for each locus.}

\item{precision_sigma}{The standard deviation of the proposal 
jump distribution for the sample inverse dispersion parameter. 
Default: \code{50000}.}

\item{mean_sigma}{The standard deviation of the proposal jump
distribution for the sample inverse dispersion parameter. 
Default: \code{0.0001}.}
}
\value{
The output is a list with 1) a vector with the mean 
of the proportion values (\code{mean_means}), 2) MAP estimate of 
the proportion values (\code{map_means}), 3) A list with count and 
loci data for the samples of interest (\code{run_data})
}
\description{
This is usually run after you have filtered the loci with zero counts 
(using \code{\link{choose_loci}} function).  It is important that, for each locus, 
at least one control sample has a non-zero count.
}
\examples{
sample_types = data.frame(Samples=c("sample_1", "sample_2", "sample_3"), Control=c(1, 1, 0)) # in practice you'll want more samples.
fit_data = fit_controls(analysis_path = "/path/to/analysis", filtered_counts, sample_types)
}
