% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_cytobands}
\alias{get_cytobands}
\title{Get cytoband data from UCSC}
\usage{
get_cytobands(genome = "hg38", chr_order = paste0("chr", c(1:22, "X",
  "Y")))
}
\arguments{
\item{genome}{The reference genome (e.g. \code{"hg38"} or \code{"hg19"}).  Default is \code{"hg38"}.}

\item{chr_order}{The chromosomes of interest and order in which to output the chromosomes.  Default is \code{paste0("chr", c(1:22, "X", "Y"))}.}
}
\value{
A tibble with cytoband data
}
\description{
Get cytoband data from UCSC
}
\examples{
cytoband = get_cytobands("hg19", c("chr1", "chr2"))
cytoband = get_cytobands() # hg38 and chromosomes 1-22, X and Y.

}
