% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loci_counts}
\alias{loci_counts}
\title{FAST-SeqS aligned read counts for OAC, GAC, BO and control samples}
\format{A data frame with 215 columns (of which 212 are samples) and 58725 rows representing genomic loci

\describe{
  \item{chr}{Chromosome}
  \item{leftPos}{The genomic coordinate within `chr`}
  \item{strand}{The strand, '+' or '-'}
}}
\usage{
loci_counts
}
\description{
A dataset containing the FAST-SeqS aligned read counts for oesophageal adenocarcinoma, 
Barrett's oesophagus, gastric adenocarcinoma and control samples.  These are
the counts obtained by the pipeline described in Abujudeh et al. 2018. Low-cost and clinically applicable
copy number profiling using repeat DNA. bioRxiv doi:
}
\references{
Abujudeh, S. et al. (2018) Low-cost and clinically applicable
copy number profiling using repeat DNA. \emph{bioRxiv doi: }
}
\keyword{datasets}
