% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{process_MCMC}
\alias{process_MCMC}
\title{Process and summarise the MCMC output from \code{\link{run_MCMC}}}
\usage{
process_MCMC(analysis_path, sample_name, fit_data, cytoband, burn_in,
  MCMC_output_folder, results_output_folder)
}
\arguments{
\item{analysis_path}{The path to the directory in 
which you want the results to be saved.
If the path does not exist, it will be created for you.}

\item{sample_name}{The sample name for which you want to obtain results}

\item{fit_data}{The \code{fit_data} list produced by \code{\link{fit_controls}}}

\item{cytoband}{Cytoband data frame produced by 
\code{\link{get_cytobands}} or \code{\link{read_cytobands}} functions}

\item{burn_in}{The first iterations to burn from the MCMC chain (Default: 5000).
For example, if the MCMC was run for 50,000 \code{iterations}
 with a \code{thin} value of 10, a \code{burn_in} of 5000 would 
 discard the first 5000 iterations (500 iterations aftering thinning)
 leaving you with 45000 iterations (4500 after thinning)}

\item{MCMC_output_folder}{The directory relative to \code{analysis_path} 
where the MCMC data was output by \code{\link{run_MCMC}}.  
The default is `MCMC`` and we recommend not to change this.}

\item{results_output_folder}{The folder, relative to \code{analysis_path} in
which the summarised output and plots will be saved.  Default is "results"
and we recommend not to change this.}
}
\value{
Returns \code{NULL} and creates some basic plots 
and a \code{sample_name.tsv} file
in \code{analysis_path/results_output_folder/sample_name/}.
}
\description{
This function process the output from \code{\link{run_MCMC}}.
It produces standard plots along with the relative copy number profile 
and credible intervals for \code{sample_name}.  It outputs this as a \code{tsv}
file in \code{analysis_path/results_output_folder/sample_name/sample_name.tsv}.
}
\examples{
process_MCMC("/path/to/analysis/",
"sample_1", fit_data, cytoband, 5000)

# see \\link{https://github.com/samabs/conliga} for more information

}
