% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{read_cytobands}
\alias{read_cytobands}
\title{Read in cytoband data (from cytoBandIdeo.txt file downloaded from UCSC)}
\usage{
read_cytobands(cytoband_file, chr_order = paste0("chr", c(1:22, "X",
  "Y")))
}
\arguments{
\item{cytoband_file}{cytoBandIdeo.txt file}

\item{chr_order}{The chromosomes of interest and order in which to output the chromosomes.  Default is \code{paste0("chr", c(1:22, "X", "Y"))}.}
}
\value{
A tibble with cytoband data
}
\description{
Read in cytoband data (from cytoBandIdeo.txt file downloaded from UCSC)
}
\examples{
cytoband = read_cytobands("path/to/cytoBandIdeo.txt", c("chr1", "chr2"))
cytoband = read_cytobands("path/to/cytoBandIdeo.txt") # chromosomes 1-22, X and Y.

}
