% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_MCMC}
\alias{run_MCMC}
\title{Run Markov Chain Monte Carlo (MCMC) to infer RCN profiles}
\usage{
run_MCMC(analysis_path, fit_data, cytoband, priors, samples, suffix,
  num_cores, output_folder, iterations, thin, gamma, rho, alpha_plus_kappa,
  max_states, init_states)
}
\arguments{
\item{analysis_path}{The path to the directory in 
which you want the results to be saved.
If the path does not exist, it will be created for you.}

\item{fit_data}{The fit_data list created by the \code{\link{fit_controls}} function}

\item{cytoband}{Cytoband information provided by the 
\code{\link{get_cytobands}} or \code{\link{read_cytobands}} functions}

\item{priors}{A data frame with the profiles defined.
If not provided, defaults will be used.}

\item{samples}{A vector of sample names you 
wish to infer RCN profiles for}

\item{suffix}{Optional: if you want to add a suffix
to the sample name}

\item{num_cores}{The number of cores to use. Default: 1.
If more than one core is used, the \code{mcmapply} function
from the \code{parallel} package
is used.}

\item{output_folder}{The output folder for the MCMC chains.
Default is `MCMC`.}

\item{iterations}{The number of iterations to the run the MCMC for
(Default: \code{50000})}

\item{thin}{The thinning of the chain
(Default: \code{10})}

\item{gamma}{The hyperparameter gamma (must be > 0).
Default: \code{1}.
If you prefer to infer gamma from the data 
(not recommended!) set it to \code{-1}.}

\item{rho}{The hyperparameter rho (value between 0-1). 
This controls how "sticky" the hidden Markov chain is.
Sensible values are around \code{0.99}. Default: \code{-1}, 
meaning it is inferred from the data.}

\item{alpha_plus_kappa}{The hyperparameter 
alpha_plus_kappa (must be > 0).
This controls the variability of the transition matrix
The greater the value, the less variance between 
rows of the transition matrix
Default: \code{-1}, meaning its value is inferred from the data}

\item{max_states}{The value to use to truncate 
the Dirichlet Process.
The default is \code{30}.  Do not set this too low.
It must be greater than the number of expected 
copy number states,
otherwise the Dirichlet approximation to the Dirichlet 
Process will not hold.}

\item{init_states}{Boolean: \code{TRUE}/\code{FALSE}.
If \code{TRUE}, 1d k-means clustering will be performed 
prior to running the chain (using the \code{Ckmeans.1d.dp}
package). This ensures the chain starts in a reasonable place.
If \code{FALSE}, the chain will be started in a random starting point.
Note by choosing \code{FALSE}, there is greater risk 
that the MCMC will fail in the first few iterations 
of the MCMC due to starting in a very unlikely position.
The default is \code{TRUE}.}
}
\value{
Returns \code{NULL} and creates MCMC output files 
for each sample in \code{analysis_path/output_folder/}.
}
\description{
This function runs the MCMC algorithm and outputs files
with samples of the posterior distribution for:
loci hidden states, state RCN, 
sample precision/inverse dispersion parameter
and other latent variables in the conliga model. This function should
be run after \code{\link{choose_loci}} and \code{\link{fit_controls}}.
Use \code{\link{process_MCMC}} to process the results of
\code{\link{run_MCMC}}.
}
\examples{
run_MCMC(analysis_path=analysis_path,
fit_data=fit_data,
cytoband=cytoband,
priors=priors,
samples=samples_to_run,
num_cores=num_cores,
iterations=50000,
thin=10,
gamma = 1, # define gamma set to 1,
max_states=30,
init_states=TRUE) # start chain off by k-means clustering

# see \\link{https://github.com/samabs/conliga} for more information
}
